requirements.txt
click==8.0.1
docopt==0.6.2
Flask==2.0.1
fpdf==1.7.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
num2words==0.5.10
Pillow==8.3.1
Werkzeug==2.0.1

marshmallow~=3.15.0